# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NextAI API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8088
security:
  - ApiKeyAuth: []
paths:
...
  /channels/qq/inbound:
    post:
      summary: Accept QQ inbound event and dispatch to agent process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: accepted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...